Navigation Distance update
Navigation system updates when user location updates. Need to update user location asynchronously

+++ app/src/main/java/r/seobo/test/UserLocation.java
@@ -1,14 +1,14 @@
 package r.seobo.test;
 
 public class UserLocation{
-    private double[] location;
+    private int[] location;
     private ChangeListener listener;
 
-    public void setLocation(double[] l){
+    public void setLocation(int[] l){
         this.location = l;
         if (listener != null) listener.onChange();
     }
-    public double[] getLocation(){
+    public int[] getLocation(){
         return this.location;
     }
     public ChangeListener getListener() {
